simplify lintrans driver
